+++ R/up_brewer.R
@@ -65,19 +65,7 @@
 #'
 #' @export
 up_brewer <- function(pik, eps = 1e-06) {
-  if (any(is.na(pik))) {
-    stop("there are missing values in the pik vector", call. = FALSE)
-  }
-  if (!is.numeric(pik)) {
-    stop("pik must be a numeric vector", call. = FALSE)
-  }
-  if (length(pik) == 0) {
-    stop("pik vector is empty", call. = FALSE)
-  }
-
-  if (any(pik < 0 | pik > 1)) {
-    stop("inclusion probabilities must be between 0 and 1", call. = FALSE)
-  }
+  check_pik(pik)
 
   .Call(C_up_brewer, as.double(pik), as.double(eps[1]))
 }